# services/booking-service/app/schemas/common.py
"""Common transport envelopes."""

from typing import Any

from pydantic import BaseModel, ConfigDict, Field


class ErrorDetail(BaseModel):
    model_config = ConfigDict(extra="forbid")

    code: str
    message: str
    retryable: bool
    details: dict[str, Any] = Field(default_factory=dict)


class ErrorEnvelope(BaseModel):
    model_config = ConfigDict(extra="forbid")

    correlationId: str
    error: ErrorDetail


class HealthResponse(BaseModel):
    service: str
    status: str
    version: str